+++ source/production/F0.Generators/CodeAnalysis/FriendlyNameGenerator.FullNameOf.cs
@@ -1,7 +1,6 @@
 using System.CodeDom.Compiler;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Threading;
 using F0.CodeDom.Compiler;
 using F0.Text;
@@ -16,18 +15,19 @@ namespace F0.CodeAnalysis
 		internal const string FullNameOf_FieldName = "fullNameOf";
 		internal const string FullNameOf_MethodName = "FullNameOf";
 
-		private static readonly SymbolDisplayFormat format = CreateFormat();
+		private static readonly SymbolDisplayFormat fullNameOfFormat = Create_FullNameOf_Format();
 
-		private static IReadOnlyCollection<FriendlyFullNameOf> GetDistinctFullNameOfInvocations(IReadOnlyList<TypeSyntax> syntaxes, Compilation compilation, CancellationToken cancellationToken)
+		[System.Diagnostics.CodeAnalysis.SuppressMessage("MicrosoftCodeAnalysisCorrectness", "RS1024:Compare symbols correctly", Justification = "https://github.com/dotnet/roslyn-analyzers/issues/4568")]
+		private static IReadOnlyCollection<ITypeSymbol> GetDistinct_FullNameOf_Invocations(IReadOnlyCollection<TypeSyntax> syntaxes, Compilation compilation, CancellationToken cancellationToken)
 		{
-			HashSet<FriendlyFullNameOf> types = new(FriendlyFullNameOfEqualityComparer.Instance);
+			HashSet<ITypeSymbol> types = new(SymbolEqualityComparer.Default);
 
 			foreach (TypeSyntax syntax in syntaxes)
 			{
 				SemanticModel semanticModel = compilation.GetSemanticModel(syntax.SyntaxTree);
 
 				ITypeSymbol? type = semanticModel.GetTypeInfo(syntax, cancellationToken).Type;
-				Debug.Assert(type is not null);
+				Debug.Assert(type is not null, $"Expression does not have a type: {syntax}");
 				Debug.Assert(type is not IErrorTypeSymbol, $"Type could not be determined due to an error: {type}");
 
 				if (type.IsTupleType)
@@ -38,13 +38,13 @@ namespace F0.CodeAnalysis
 					type = namedType.TupleUnderlyingType ?? type;
 				}
 
-				_ = types.Add(new FriendlyFullNameOf(type));
+				_ = types.Add(type);
 			}
 
 			return types;
 		}
 
-		private static void Write_FullNameOf_FieldDeclaration_To(IndentedTextWriter writer, IReadOnlyCollection<FriendlyFullNameOf> types, LanguageFeatures features)
+		private static void Write_FullNameOf_FieldDeclaration_To(IndentedTextWriter writer, IReadOnlyCollection<ITypeSymbol> types, LanguageFeatures features)
 		{
 			if (features.HasGenerics && types.Count > 0)
 			{
@@ -52,7 +52,7 @@ namespace F0.CodeAnalysis
 			}
 		}
 
-		private static void Write_FullNameOf_MethodDeclaration_To(IndentedTextWriter writer, IReadOnlyCollection<FriendlyFullNameOf> types, LanguageFeatures features)
+		private static void Write_FullNameOf_MethodDeclaration_To(IndentedTextWriter writer, IReadOnlyCollection<ITypeSymbol> types, LanguageFeatures features)
 		{
 			if (features.HasGenerics)
 			{
@@ -83,7 +83,7 @@ namespace F0.CodeAnalysis
 			writer.WriteLine(Tokens.CloseBrace);
 		}
 
-		private static void Write_CreateFullNameOf_MethodDeclaration_To(IndentedTextWriter writer, IReadOnlyCollection<FriendlyFullNameOf> types, LanguageFeatures features)
+		private static void Write_CreateFullNameOf_MethodDeclaration_To(IndentedTextWriter writer, IReadOnlyCollection<ITypeSymbol> types, LanguageFeatures features)
 		{
 			if (features.HasGenerics && types.Count > 0)
 			{
@@ -106,20 +106,20 @@ namespace F0.CodeAnalysis
 				writer.WriteLine($"{Tokens.CloseBrace}");
 			}
 
-			static void ObjectCreationExpression(IndentedTextWriter writer, IReadOnlyCollection<FriendlyFullNameOf> types, LanguageFeatures features)
+			static void ObjectCreationExpression(IndentedTextWriter writer, IReadOnlyCollection<ITypeSymbol> types, LanguageFeatures features)
 			{
 				Debug.Assert(!features.HasImplicitlyTypedLocalVariable);
 				writer.WriteLine($"global::System.Collections.Generic.Dictionary<global::System.Type, string> dictionary = new global::System.Collections.Generic.Dictionary<global::System.Type, string>({types.Count});");
 
-				foreach (FriendlyFullNameOf type in types)
+				foreach (ITypeSymbol type in types)
 				{
-					writer.WriteLine($@"dictionary.Add(typeof({type.Symbol.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)}), ""{type.Symbol.ToDisplayString(format)}"");");
+					writer.WriteLine($@"dictionary.Add(typeof({type.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)}), ""{type.ToDisplayString(fullNameOfFormat)}"");");
 				}
 
 				writer.WriteLine($"return dictionary;");
 			}
 
-			static void CollectionInitializerExpression(IndentedTextWriter writer, IReadOnlyCollection<FriendlyFullNameOf> types, LanguageFeatures features)
+			static void CollectionInitializerExpression(IndentedTextWriter writer, IReadOnlyCollection<ITypeSymbol> types, LanguageFeatures features)
 			{
 				if (features.HasTargetTypedObjectCreation)
 				{
@@ -132,46 +132,17 @@ namespace F0.CodeAnalysis
 				writer.WriteLine(Tokens.OpenBrace);
 				writer.Indent++;
 
-				foreach (FriendlyFullNameOf type in types)
+				foreach (ITypeSymbol type in types)
 				{
-					writer.WriteLine($@"{{ typeof({type.Symbol.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)}), ""{type.Symbol.ToDisplayString(format)}"" }},");
+					writer.WriteLine($@"{{ typeof({type.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)}), ""{type.ToDisplayString(fullNameOfFormat)}"" }},");
 				}
 
 				writer.Indent--;
-				writer.WriteLine($"{Tokens.CloseBrace};");
+				writer.WriteLine($"{Tokens.CloseBrace}{Tokens.Semicolon}");
 			}
 		}
 
-		private sealed class FriendlyFullNameOf
-		{
-			public FriendlyFullNameOf(ITypeSymbol symbol)
-			{
-				Symbol = symbol;
-			}
-
-			public ITypeSymbol Symbol { get; }
-		}
-
-		private sealed class FriendlyFullNameOfEqualityComparer : EqualityComparer<FriendlyFullNameOf>
-		{
-			public static IEqualityComparer<FriendlyFullNameOf> Instance { get; } = new FriendlyFullNameOfEqualityComparer();
-
-			private FriendlyFullNameOfEqualityComparer()
-			{
-			}
-
-			public override bool Equals(FriendlyFullNameOf? x, FriendlyFullNameOf? y)
-			{
-				return SymbolEqualityComparer.Default.Equals(x?.Symbol, y?.Symbol);
-			}
-
-			public override int GetHashCode([DisallowNull] FriendlyFullNameOf obj)
-			{
-				return SymbolEqualityComparer.Default.GetHashCode(obj.Symbol);
-			}
-		}
-
-		private static SymbolDisplayFormat CreateFormat()
+		private static SymbolDisplayFormat Create_FullNameOf_Format()
 		{
 			SymbolDisplayTypeQualificationStyle typeQualificationStyle = SymbolDisplayTypeQualificationStyle.NameAndContainingTypesAndNamespaces;