+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Fixed `FriendlyNameGenerator`, now generating `Friendly.NameOf<T>()` to always return a simple name (and the containing types), rather than including namespaces when invoked with a qualified name without a using directive in scope of the call site.
 
 ## v0.1.0 (2021-07-01)
 - Added `FriendlyNameGenerator`, which generates both `Friendly.NameOf<T>()` and `Friendly.FullNameOf<T>()`, pretty printing (fully qualified) type names, backed by a reflection-free lookup.